changes to old product routes, auth middleware, controller

+++ src/controller/product.controller.js
@@ -10,7 +10,7 @@ const { v4: uuidv4 } = require('uuid')
 const cloudinary = require("../config/cloudinary")
 
 //Imporing file system library 
-const fs = require("fs"); 
+const fs = require("fs");
 
 //Get all products async function 
 const get = asyncHandler(async (req, res) => {
@@ -32,7 +32,7 @@ const show = asyncHandler(async (req, res) => {
     const { id } = req.params
 
     //Fetching single product using the id in the req.params from the database and assigning it to product
-    const product = await productModel.findOne({ productId: id });
+    const product = await productModel.find({ productId: id });
 
     try {
         if (product) {
@@ -61,13 +61,16 @@ const create = asyncHandler(async (req, res) => {
     try {
 
         if (req.method === 'POST') {
+
             const urls = []
-            const files = req.files;
-            for (const file of files) {
-                const { path } = file;
-                const newPath = await uploader(path)
-                urls.push(newPath)
-                fs.unlinkSync(path)
+            if (req.files) {
+                const files = req.files;
+                for (const file of files) {
+                    const { path } = file;
+                    const newPath = await uploader(path)
+                    urls.push(newPath)
+                    fs.unlinkSync(path)
+                }
             }
 
             //we use uuidv4 to generate a random and unique id for the products
@@ -83,7 +86,7 @@ const create = asyncHandler(async (req, res) => {
             })
 
             product.save()
-            return res.status(200).json({
+            return res.status(201).json({
                 success: true,
                 message: "product created sucessfully",
                 data: product
@@ -95,9 +98,9 @@ const create = asyncHandler(async (req, res) => {
         }
 
     } catch (error) {
-        return res.status(412).send({
+        return res.status(412).json({
             success: false,
-            message: error.message
+            message: error
         })
     }
 
@@ -105,58 +108,49 @@ const create = asyncHandler(async (req, res) => {
 
 const update = asyncHandler(async (req, res) => {
     //Destruct the data sent from req.body 
-    const { name, description, price } = req.body
-
     const uploader = async (path) => await cloudinary.uploads(path, "Images")
 
     //Destructing the id from req.params
     const { id } = req.params
-
     //assigning the specfic product to variable called product
-    let product = await productModel.findOne({ productId: id });
+    const product = await productModel.findOne({ productId: id });
+
+    console.log("id", id)
+    if (req.method !== 'PUT') {
+        return res.status(405).json({
+            err: `${req.method} method not allowed`
+        })
+    }
 
     try {
 
-        if (product) {
-            //updating the datas of that product
-            if (req.method === 'PUT') {
-                const urls = []
-                const files = req.files;
-                for (const file of files) {
-                    const { path } = file;
-                    const newPath = await uploader(path)
-                    urls.push(newPath)
-                    fs.unlinkSync(path)
-                }
-
-                //Updating the product
-                product.updateOne(
-                    {
-                        $set: {
-                            name: name,
-                            description: description,
-                            price: price,
-                        }
-                    },
-                    {}, { new: true }
-                )
-
-                return res.status(200).json({
-                    success: true,
-                    message: "product updated sucessfully",
-                    data: req.body
-                })
-            } else {
-                return res.status(405).json({
-                    err: `${req.method} method not allowed`
-                })
-            }
-        } else {
+        if (!product) {
             return res.status(400).json({
                 success: false,
                 message: "product not found",
             })
         }
+
+        //updating the datas of that product
+        const urls = []
+        if (req.files) {
+            const files = req.files;
+            for (const file of files) {
+                const { path } = file;
+                const newPath = await uploader(path)
+                urls.push(newPath)
+                fs.unlinkSync(path)
+            }
+
+        }
+
+        product.updateOne(req.body, { useFindAndModify: false }).then(data => {
+            if (!data) {
+                res.status(404).send({
+                    message: `Cannot update product with id=${id}. Maybe product was not found!`
+                });
+            } else return res.status(201).json({ message: "Product was updated successfully.", data: req.body });
+        })
 
     } catch (error) {
         return res.status(412).send({
@@ -177,7 +171,7 @@ const deleteProduct = asyncHandler(async (req, res) => {
         await productModel.deleteOne({ productId: id });
 
         //Since there is no data to be responde we simple send a message
-        return res.status(200).json({
+        return res.status(410).json({
             success: true,
             message: "product deleted sucessfully",
         })
